requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Medical imaging
nibabel>=5.0.0
pydicom>=2.3.0

# Image processing
scikit-image>=0.20.0
Pillow>=9.4.0

# Visualization and logging
wandb>=0.15.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
